backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-colorfield
django-cors-headers
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg
gunicorn==20.1.0
idna==3.3
Pillow==8.4.0
psycopg2-binary==2.9.1
PyJWT==2.3.0
python-dotenv==0.19.1
pytils==0.3
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
